requirements.txt
streamlit
streamlit-agraph
dashscope
pandas
networkx
Pillow
numpy
